DELIA-21705 : Change the variable type to long long to support big numbers.
Reason for change: we used 'unsigned long' to report/return the capacity n freespace of the memory devices but noticed that this variable is 4 bytes size but the XG platforms has big sizes n the value must be reported in 8 byte length variable to represent properly. Fixed it now. This needs a API change, the XRE/ServiceManager, HostIf, SNMP will be modified to fix this.
Test Procedure: Please verify this in XG platforms
Risks: Medium

+++ refactored/include/rdkStorageMgr.h
@@ -36,16 +36,16 @@ eSTMGRReturns rdkStorage_getTSBMaxMinutes (unsigned int *pMinutes);
 eSTMGRReturns rdkStorage_getTSBCapacityMinutes(unsigned int *pMinutes);
 
 /* Get TSBCapacity*/
-eSTMGRReturns rdkStorage_getTSBCapacity(unsigned long *pCapacity);
+eSTMGRReturns rdkStorage_getTSBCapacity(unsigned long long *pCapacity);
 
 /* Get TSBFreeSpace*/
-eSTMGRReturns rdkStorage_getTSBFreeSpace(unsigned long *pFreeSpace);
+eSTMGRReturns rdkStorage_getTSBFreeSpace(unsigned long long *pFreeSpace);
 
 /* Get DVRCapacity */
-eSTMGRReturns rdkStorage_getDVRCapacity(unsigned long *pCapacity);
+eSTMGRReturns rdkStorage_getDVRCapacity(unsigned long long *pCapacity);
 
 /* Get DVRFreeSpace*/
-eSTMGRReturns rdkStorage_getDVRFreeSpace(unsigned long *pFreeSpace);
+eSTMGRReturns rdkStorage_getDVRFreeSpace(unsigned long long *pFreeSpace);
 
 /* Get isTSBEnabled*/
 bool rdkStorage_isTSBEnabled();
