+++ refactored/mgr/rdkStorage.cpp
@@ -411,7 +411,7 @@ eSTMGRReturns rdkStorage_getTSBCapacityMinutes(unsigned int *pMinutes)
 }
 
 /* Get TSBCapacity*/
-eSTMGRReturns rdkStorage_getTSBCapacity(unsigned long *pCapacity)
+eSTMGRReturns rdkStorage_getTSBCapacity(unsigned long long *pCapacity)
 {
     STMGRLOG_DEBUG("ENTRY of %s\n", __FUNCTION__);
 
@@ -419,7 +419,7 @@ eSTMGRReturns rdkStorage_getTSBCapacity(unsigned long *pCapacity)
 }
 
 /* Get TSBFreeSpace*/
-eSTMGRReturns rdkStorage_getTSBFreeSpace(unsigned long *pFreeSpace)
+eSTMGRReturns rdkStorage_getTSBFreeSpace(unsigned long long *pFreeSpace)
 {
     STMGRLOG_DEBUG("ENTRY of %s\n", __FUNCTION__);
 
@@ -427,7 +427,7 @@ eSTMGRReturns rdkStorage_getTSBFreeSpace(unsigned long *pFreeSpace)
 }
 
 /* Get DVRCapacity */
-eSTMGRReturns rdkStorage_getDVRCapacity(unsigned long *pCapacity)
+eSTMGRReturns rdkStorage_getDVRCapacity(unsigned long long *pCapacity)
 {
     STMGRLOG_DEBUG("ENTRY of %s\n", __FUNCTION__);
 
@@ -435,7 +435,7 @@ eSTMGRReturns rdkStorage_getDVRCapacity(unsigned long *pCapacity)
 }
 
 /* Get DVRFreeSpace*/
-eSTMGRReturns rdkStorage_getDVRFreeSpace(unsigned long *pFreeSpace)
+eSTMGRReturns rdkStorage_getDVRFreeSpace(unsigned long long *pFreeSpace)
 {
     STMGRLOG_DEBUG("ENTRY of %s\n", __FUNCTION__);
